requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.4.0
redis==5.0.4
psycopg2-binary==2.9.9
flower==2.0.1
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0
apscheduler==3.10.4
openai>=1.30.0
paramiko==3.4.0
google-auth
